SUPERNOVA: Fix Remaining GCC Warnings.
These are more difficult and are likely logic errors i.e. "comparison of
constant ‘15’ with boolean expression is always true [-Wbool-compare]"
so will require further work, but this removes the warning for now and
shouldn't change the existing logical behaviour.

Several FIXME comments are added to mark the problematic statements.