IMMORTAL: Adds palette fading to Renderer
The fading functions block execution during the fading thus cannot react
on user input or update the logic state. As of now, this doesn't pose a
problem but may in the future.
The reason for splitting the member functions like drawImage() and
internalDrawImage() is to provide a public interface for calling those
functions and internally handle them, like deferring them to the next
update() call.