IMMORTAL: Handles YesNo Dialog token
As the TODO says, this is just a workaround right now and I haven't
found the reason for the misplaced buttons yet.
According to the source only the life meters in the fight scenes add
the sprite x position to the absolute position in Renderer::drawSprite()
but adding them for text as well fixed them as well as animated sprites
(otherwise they were moving in uninteded ways if the width of the shortest
transparent line was not identical to the previous sprite).