IMMORTAL: Removes Animations
Instead of packing sprites to animations while parsing, the resource
manager now stores the sprites on their own. The reason is simple, not
every sprite is part of an animation and most do not consist of a range
of consequtive sprites, so defining them got quite tedious.

This commit breaks the sprites for the YesNo dialog token thus the
drawSprite calls in runDialog() are commented out.